% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disweights.R
\docType{data}
\name{disweights}
\alias{disweights}
\title{Dataset: Health Metrics Data - Disability Weights and Severity in 2019 and 2021}
\format{
A dataframe with 463 rows and 9 variables:

The variables are as follows:

\describe{
\item{sequela}{character, disease sequela}
\item{specification}{character, diesase specification}
\item{cause1}{character, first cause of disease - morbidity}
\item{cause2}{character, second cause of disease - morbidity}
\item{severity}{character, mild,
moderate,
severe,
mean}
\item{dw}{numeric, disability weights estimation}
\item{upper}{numeric, upper value estimation}
\item{lower}{numeric, lower value estimation}
}
}
\source{
{Global Burden of Disease Collaborative Network. Global Burden of Disease Study 2019 and 2021 Disability Weights. Seattle, United States of America: Institute for Health Metrics and Evaluation (IHME), 2024.}
}
\usage{
disweights
}
\description{
A dataset containing the disability weights estimates, upper and lower values,
and the severity level for Stroke, Tuberculosis, and HIV for all countries.
}
\keyword{datasets}
