% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disweights2019.R
\docType{data}
\name{disweights2019}
\alias{disweights2019}
\title{Health Metrics - Disability Weights and Severity in 2019}
\format{
A dataframe with 25 rows and 7 variables:

The variables are as follows:

\describe{
\item{specification}{character, diesase specification}
\item{cause1}{character, first cause of disease}
\item{cause2}{character, second cause of disease}
\item{severity}{character, mild,
moderate,
severe}
\item{disability_weight}{numeric, disability weight estimation}
\item{upper}{numeric, upper value estimation}
\item{lower}{numeric, lower value estimation}
}
}
\source{
{2019 data from the \href{https://vizhub.healthdata.org/gbd-results/}{IHME} website}
}
\usage{
disweights2019
}
\description{
A dataset containing the disability weights estimates, upper and lower values,
and the severity level for Stroke, Tuberculosis, and HIV for 2019.
}
\keyword{datasets}
