% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{fetch_data}
\alias{fetch_data}
\title{Helper function to fetch data from API}
\usage{
fetch_data(url, key, endpoint, query_params)
}
\description{
This is an internal function used to fetch data from the specified API.
It is not exported and should not be called directly by users.
}
\keyword{internal}
