% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g7_inc_prev_CVD2019.R
\docType{data}
\name{g7_inc_prev_CVD2019}
\alias{g7_inc_prev_CVD2019}
\title{Dataset: Health Metrics Data - G7 Countries Incidence and Prevalence for CVD and Stroke in 2019}
\format{
A dataframe with 1824 rows and 8 variables:

The variables are as follows:

\describe{
\item{measure}{character, metric name}
\item{location}{character, country}
\item{sex}{character, Female, Male, Both}
\item{age}{character, age groups from 01 to 85+ each 5 years}
\item{cause}{character, CVD and Stroke}
\item{val}{numeric, estimated values}
\item{upper}{numeric, estimated upper values}
\item{lower}{numeric, estimated lower values}
}
}
\source{
\url{https://vizhub.healthdata.org/gbd-results/}
}
\usage{
g7_inc_prev_CVD2019
}
\description{
Locations available are Global, Canada, France, Germany, Italy, Japan, UK, and US.
}
\keyword{datasets}
