% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_get_data.R
\name{gbd_get_data}
\alias{gbd_get_data}
\title{Title: gbd_get_data - Fetch Data from GBD API}
\usage{
gbd_get_data(url, key, endpoint, ...)
}
\arguments{
\item{url}{The base URL of the API.}

\item{key}{The API key for authorization.}

\item{endpoint}{The specific endpoint to retrieve data from.}

\item{...}{Additional query parameters such as location_id, year, etc.}
}
\value{
A data frame or list of results from the API.
}
\description{
This function fetches data from the GBD API. To use this function, you need to have an API key. You can get the key by registering on the \href{https://ghdx.healthdata.org/ihme-api}{IHME-API} website.
}
\examples{
\dontrun{
# This is a dontrun example because it requires an API KEY.
url <- "https://api.healthdata.org/sdg/v1"
key <- "YOUR-KEY"
endpoint <- "GetResultsByIndicator"

data <- gbd_get_data(url,
                    key,
                    endpoint,
                    indicator_id="1001",
                    location_id= c("29","86","102"),
                    year="2019",
                    limit = 10)
}

}
