% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germany_lungc.R
\docType{data}
\name{germany_lungc}
\alias{germany_lungc}
\title{Germany lungcancer deaths 2019}
\format{
A dataframe with 48 rows and 5 variables:

The variables are as follows:

\describe{
\item{sex}{character, both, male, female}
\item{age}{character, age groups from 10-14 to 85+ each 5 years}
\item{val}{numeric, deaths number estimation due to lungcancer}
\item{upper}{numeric, upper value estimation}
\item{lower}{numeric, lower value estimation}
}
}
\source{
{2019 data from the \href{https://vizhub.healthdata.org/gbd-results/}{IHME} website}
}
\usage{
germany_lungc
}
\description{
A dataset containing deaths number due to lungcancer in Germany 2019.
}
\keyword{datasets}
