% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gho_le_hale.R
\docType{data}
\name{gho_le_hale}
\alias{gho_le_hale}
\title{Global Health Observatory (GHO) - Countries Life Expectancy and HALE 2000-2019}
\format{
A dataframe with 8784 rows and 6 variables:

The variables are as follows:

\describe{
\item{indicator}{character, Healthy life expectancy (HALE) at age 60 (years),\if{html}{\out{<br>}}
Healthy life expectancy (HALE) at birth (years),\if{html}{\out{<br>}}
Life expectancy at age 60 (years),\if{html}{\out{<br>}}
Life expectancy at birth (years)}
\item{year}{numeric, from 2000 to 2019}
\item{region}{character, 6 World regions: Africa, Americas, Eastern Mediterranean, Europe, South-East Asia, and Western Pacific}
\item{country}{character, 183 World countries}
\item{sex}{character, both, male, female}
\item{value}{numeric, value of the indicator}
}
}
\source{
{\href{https://apps.who.int/gho/data/node.main.SDG2016LEX?lang=en}{WHO}}
}
\usage{
gho_le_hale
}
\description{
A dataset containing World countries Life Expectancy and HALE from 2000 to 2019.
}
\keyword{datasets}
