% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gho_lifetables.R
\docType{data}
\name{gho_lifetables}
\alias{gho_lifetables}
\title{Global Health Observatory (GHO) life tables: WHO Global life table values}
\format{
A dataframe with 1995 rows and 5 variables:

Variables:

\describe{
\item{indicator}{Tx - person-years lived above age x,\if{html}{\out{<br>}}
ex - expectation of life at age x,\if{html}{\out{<br>}}
lx - number of people left alive at age x,\if{html}{\out{<br>}}
nLx - person-years lived between ages x and x+n,\if{html}{\out{<br>}}
nMx - age-specific death rate between ages x and x+n,\if{html}{\out{<br>}}
ndx - number of people dying between ages x and x+n,\if{html}{\out{<br>}}
nqx - probability of dying between ages x and x+n}
\item{year}{from 2000 to 2019}
\item{age}{from <1 to 85+ each 5 years}
\item{sex}{both, male, female}
\item{value}{value of the tables}
}
}
\source{
{\href{https://apps.who.int/gho/data/node.main.LIFECOUNTRY?lang=en}{WHO}}
}
\usage{
gho_lifetables
}
\description{
A dataset containing the Global region life tables from 2000 to 2019
The variables are as follows:
}
\keyword{datasets}
