% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmsidwR-package.R
\docType{package}
\name{hmsidwR-package}
\alias{hmsidwR}
\alias{hmsidwR-package}
\title{hmsidwR: Health Metrics and the Spread of Infectious Diseases}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A collection of datasets and supporting functions accompanying Health Metrics and the Spread of Infectious Diseases by Federica Gazzelloni (2024). This package provides data for health metrics calculations, including Disability-Adjusted Life Years (DALYs), Years of Life Lost (YLLs), and Years Lived with Disability (YLDs), as well as additional tools for analyzing and visualizing health data. Federica Gazzelloni (2024) \doi{10.5281/zenodo.10818338}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Fgazzelloni/hmsidwR}
  \item \url{https://fgazzelloni.github.io/hmsidwR/}
  \item Report bugs at \url{https://github.com/Fgazzelloni/hmsidwR/issues}
}

}
\author{
\strong{Maintainer}: Federica Gazzelloni \email{fede.gazzelloni@gmail.com} (\href{https://orcid.org/0000-0002-4285-611X}{ORCID})

}
\keyword{internal}
