% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incprev_stroke.R
\docType{data}
\name{incprev_stroke}
\alias{incprev_stroke}
\title{Global Region Health Metrics Data - Incidence and Prevalence for Stroke 2019 and 2021
Numbers - 5-year age groups from <1 to 85+ and both
Location available Global}
\format{
A dataframe with 228 rows and 7 variables:

The variables are as follows:

\describe{
\item{measure}{character, metric name}
\item{sex}{character, female, male, both}
\item{age}{character, age groups from <1 to 85+ each 5 years}
\item{year}{integer, years 2019 and 2021}
\item{val}{numeric, estimated values}
\item{upper}{numeric, estimated upper values}
\item{lower}{numeric, estimated lower values}
}
}
\source{
\url{https://vizhub.healthdata.org/gbd-results/}
}
\usage{
incprev_stroke
}
\description{
Global Region Health Metrics Data - Incidence and Prevalence for Stroke 2019 and 2021
Numbers - 5-year age groups from <1 to 85+ and both
Location available Global
}
\keyword{datasets}
