% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infectious_diseases.R
\docType{data}
\name{infectious_diseases}
\alias{infectious_diseases}
\title{Dataset: Health Metrics Data - Infectious Diseases 1980-2021}
\format{
A dataframe with 7470 rows and 10 variables:

The variables are as follows:

\describe{
\item{year}{numeric, from 1980 to 2021}
\item{location_name}{character, list of countries}
\item{location_id}{numeric, list of countries by id}
\item{cause_name}{character, type of infectious disease}
\item{Deaths}{numeric, deaths rate}
\item{DALYs}{numeric, DALYs for 100 000}
\item{YLDs}{numeric, YLDs for 100 000}
\item{YLLs}{numeric, YLLs for 100 000}
\item{Prevalence}{numeric, prevalence rate}
\item{Incidence}{numeric, incidence rate}
\item{val}{numeric, estimated values}
}
}
\source{
{\href{https://vizhub.healthdata.org/gbd-results/}{IHME} website}
}
\usage{
infectious_diseases
}
\description{
A dataset containing deaths rates, dalys, ylls, ylds, prevalence and incidence due to infectious diseases form 1980 to 2021 for Lesotho, Eswatini, Malawi, Central African Republic, and Zambia.
}
\keyword{datasets}
