% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmforecast.R
\name{lmforecast}
\alias{lmforecast}
\title{lmforecast function}
\usage{
lmforecast(x, a, m, e)
}
\arguments{
\item{x}{A numeric vector, identifying the covariate of the model}

\item{a}{A numeric vector, The intercept}

\item{m}{A numeric vector, The slope, or the betas estimated values}

\item{e}{A numeric vector, The error}
}
\value{
A dataframe containing the vector with the estimation, and the original data
}
\description{
lmforecast function
}
\examples{
x <- rnorm(n = 50, mean = 0, sd = 1)
lmforecast(
x = x,
a = 0.1,
m = dunif(x, min = 0.02, max = 1),
e = rnorm(length(x), mean = 0, sd = 1))
}
