% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meningitis.R
\docType{data}
\name{meningitis}
\alias{meningitis}
\title{Dataset: Health Metrics Data - Meningitis}
\format{
A dataframe with 160 rows and 6 variables:

The variables are as follows:

\describe{
\item{year}{character, metric name}
\item{location}{character, country}
\item{Deaths}{numeric, Rates}
\item{DALYs}{numeric, Rates}
\item{Smoking}{numeric, Rates}
\item{Particulate matter pollution}{PM2.5, numeric, Rates}
}
}
\source{
\url{https://vizhub.healthdata.org/gbd-results/}
}
\usage{
meningitis
}
\description{
Data include the number of deaths due to meningitis, as well as levels of two risk factors, particulate matter (PM2.5) and smoking.
}
\details{
For years from 1990 to 2021.

Locations available are Central African Republic, Zambia, Eswatini, Lesotho, and Malawi.
}
\keyword{datasets}
