% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_search.R
\name{string_search}
\alias{string_search}
\title{Scan all folders and files to find a string: string_search}
\usage{
string_search(path = ".", pattern, string)
}
\arguments{
\item{path}{If NULL, the current directory is used}

\item{pattern}{A regular expression pattern such as '\\.R$'}

\item{string}{A string such as 'metric'}
}
\value{
A character vector with the names of the files that contain the string
}
\description{
Scan all folders and files to find a string: string_search
}
\examples{
string_search(path=".","\\\\.R$","metric")

# function string_search
}
