% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disweights2019.R
\docType{data}
\name{disweights2019}
\alias{disweights2019}
\title{Dataset: Health Metrics Data - Disability Weights and Severity in 2019}
\format{
A dataframe with 133 rows and 8 variables:

The variables are as follows:

\describe{
\item{sequela}{character, disease sequela}
\item{specification}{character, diesase specification}
\item{cause1}{character, first cause of disease - morbidity}
\item{cause2}{character, second cause of disease - morbidity}
\item{severity}{character, mild,
moderate,
severe,
combined}
\item{disability_weight}{numeric, disability weight estimation}
\item{upper}{numeric, upper value estimation}
\item{lower}{numeric, lower value estimation}
}
}
\source{
{Global Burden of Disease Collaborative Network. Global Burden of Disease Study 2019 (GBD 2019) Disability Weights. Seattle, United States of America: Institute for Health Metrics and Evaluation (IHME), 2020. DOI  https://doi.org/10.6069/1W19-VX76}
}
\usage{
disweights2019
}
\description{
A dataset containing the disability weights estimates, upper and lower values,
and the severity level for Stroke, Tuberculosis, and HIV for 2019.
}
\keyword{datasets}
