% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g7_hmetrics.R
\docType{data}
\name{g7_hmetrics}
\alias{g7_hmetrics}
\title{Dataset: Health Metrics Data - G7 Countries}
\format{
A dataframe with 3402 rows and 9 variables:

The variables are as follows:

\describe{
\item{measure}{character, metric name}
\item{location}{character, country}
\item{sex}{character, Female, Male, Both}
\item{cause}{character, all causes, and respiratory infections and tuberculosis}
\item{year}{integer, year}
\item{val}{numeric, estimated values}
\item{upper}{numeric, estimated upper values}
\item{lower}{numeric, estimated lower values}
}
}
\source{
\url{https://vizhub.healthdata.org/gbd-results/}
}
\usage{
g7_hmetrics
}
\description{
A subset of data from the IHME GBD on Deaths, Disability-Adjusted Life Years (DALYs), Years of Life Lost (YLLs), and Years Lived with Disability (YLDs), Incidence and Prevalence, age standardized for all causes and respiratory infections and tuberculosis.
For years 2010, 2019 and 2021.
}
\details{
Locations available are Global, Canada, France, Germany, Italy, Japan, UK, and US.
}
\keyword{datasets}
