% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getunz.R
\name{getunz}
\alias{getunz}
\title{Title}
\usage{
getunz(zipfile)
}
\arguments{
\item{zipfile}{A url string, or a .zip file stored locally.}
}
\value{
A dataframe object from a zipped file.
The function create a temp directory, unzip the file, list the available
csv files in it and read them.
}
\description{
Title
}
\examples{
\dontrun{
getunz("data.zip")
}
}
