% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getunz.R
\name{getunz}
\alias{getunz}
\title{Download, Unzip and Read Data: getunz}
\usage{
getunz(url)
}
\arguments{
\item{url}{A url string for a .zip file.}
}
\value{
A dataframe object from a zipped file.
The function read the url, create a temp directory,
unzip the file, list the available
csv files in it and read them.
}
\description{
Download, Unzip and Read Data: getunz
}
\examples{
\dontrun{
getunz("url")
}
}
