% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmsidwR-package.R
\docType{package}
\name{hmsidwR-package}
\alias{hmsidwR}
\alias{hmsidwR-package}
\title{hmsidwR: Health Metrics and the Spread of Infectious Diseases with R}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A collection of data and functions used in the Health Metrics and the Spread of Infectious Diseases with R book. The book provides the tools for calculating the DALYs (YLLs+YLDs), and other important health metrics. It also shows the R code for modeling infectious diseases with machine learning. Finally, allows the user to understand how to visualize the results and making maps.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Fgazzelloni/hmsidwR}
  \item \url{https://fgazzelloni.github.io/hmsidwR/}
  \item Report bugs at \url{https://github.com/Fgazzelloni/hmsidwR/issues}
}

}
\author{
\strong{Maintainer}: Federica Gazzelloni \email{fede.gazzelloni@gmail.com} (\href{https://orcid.org/0000-0002-4285-611X}{ORCID})

}
\keyword{internal}
