% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_affected_countries.R
\docType{data}
\name{id_affected_countries}
\alias{id_affected_countries}
\title{Health Metrics - Infectious Diseases 1980-2021}
\format{
A dataframe with 3066 rows and 6 variables:

The variables are as follows:

\describe{
\item{location_name}{character, list of countries}
\item{year}{numeric, from 1980 to 2021}
\item{DALYs}{numeric, DALYs for 100 000}
\item{YLLs}{numeric, YLLs for 100 000}
\item{YLDs}{numeric, YLDs for 100 000}
\item{Deaths}{numeric, deaths rate}
}
}
\source{
{\href{https://vizhub.healthdata.org/gbd-results/}{IHME} website}
}
\usage{
id_affected_countries
}
\description{
A dataset containing average values for deaths rates, dalys, ylls, ylds due to 37 infectious diseases form 1980 to 2012 for all countries.
}
\keyword{datasets}
