% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idDALY_map_data.R
\docType{data}
\name{idDALY_map_data}
\alias{idDALY_map_data}
\title{Health Metrics - Simple Feature Collection
Average DALYs per 100,000 population from 1990 to 2021}
\format{
A Simple feature collection with 1402 rows and 4 variables:

\describe{
\item{group}{double, country's polygon}
\item{location_name}{character, 200 Countries affected by Infectious Diseases}
\item{DALYs}{double, Average DALYs per 100,000 population from 1990 to 2021}
\item{geometry}{POLYGON}
}
}
\source{
{2021 data from the \href{https://vizhub.healthdata.org/gbd-results/}{IHME} website}
}
\usage{
idDALY_map_data
}
\description{
Health Metrics - Simple Feature Collection
Average DALYs per 100,000 population from 1990 to 2021
}
\keyword{datasets}
