% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rabies.R
\docType{data}
\name{rabies}
\alias{rabies}
\title{Dataset: Health Metrics Data - Rabies Deaths and DALYs from 1980 to 2021}
\format{
A dataframe with 296 rows and 7 variables:

The variables are as follows:

\describe{
\item{measure}{character, metric name}
\item{location}{character, country}
\item{cause}{character, cause}
\item{year}{integer, year}
\item{val}{numeric, estimated values}
\item{upper}{numeric, estimated upper values}
\item{lower}{numeric, estimated lower values}
}
}
\source{
\url{https://www.healthdata.org/}
}
\usage{
rabies
}
\description{
A subset of data from the IHME GBD on DALYs and Deaths due to All Causes and Rabies. Locations available are Global Region and Asia.
}
\keyword{datasets}
