% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdi90_19.R
\docType{data}
\name{sdi90_19}
\alias{sdi90_19}
\title{Socio-Demographic Index (SDI) for 1990 and 2019}
\format{
A dataframe with 20010 rows and 3 variables:

The variables are as follows:

\describe{
\item{location}{character, country}
\item{year}{integer, year}
\item{val}{numeric, estimated values}
}
}
\source{
<healthdata.org>
}
\usage{
sdi90_19
}
\description{
A subset of data from the IHME GBD containing location, year and
estimated values of the SDI for the years 1990 and 2019.
}
\keyword{datasets}
